change data type to long

+++ src/main/java/com/odde/massivemailer/service/TrackingService.java
@@ -2,6 +2,6 @@ package com.odde.massivemailer.service;
 
 public interface TrackingService {
 
-    int updateViewCount(int messageId, String userId);
+    int updateViewCount(long messageId, String userId);
 
 }
